+++ plugins/aws/test/resources/files/ec2/describe-instances.json
