+++ plugins/aws/test/resources/files/ec2/describe-key-pairs.json
