+++ plugins/aws/test/resources/files/ec2/describe-network-acls.json
