+++ plugins/aws/test/resources/files/ec2/describe-volumes.json
