+++ plugins/aws/test/resources/files/iam/get-access-key-last-used__ak_test
