+++ plugins/aws/test/resources/files/iam/list-access-keys__test_user.json
