+++ plugins/aws/test/resources/files/iam/list-attached-group-policies__group1.json
