+++ plugins/aws/test/resources/files/iam/list-attached-role-policies__role1.json
