+++ plugins/aws/test/resources/files/iam/list-group-policies__group1.json
