+++ plugins/aws/test/resources/files/iam/list-groups.json
