+++ plugins/aws/test/resources/files/iam/list-instance-profiles-for-role__role1.json
