+++ plugins/aws/test/resources/files/iam/list-policies.json
