+++ plugins/aws/test/resources/files/iam/list-role-policies__role1.json
