+++ plugins/aws/test/resources/files/iam/list-roles.json
