+++ plugins/aws/test/resources/files/iam/list-server-certificates.json
