+++ plugins/aws/test/resources/files/iam/list-user-policies__test_user.json
